Created a new file called goals.uility to store all the functions that have to do with goals. Did some refactoring with the new changes.
